Move `HGraph` to its own file.
Ideally we would not need to `#include "graph.h"` in the
`nodes.h` but that would require extreme refactoring. More
realistic aim is to include `graph.h` instead of the whole
`nodes.h` for some compilation units.

Note that `HGraph::AllocateInstructionId()` is no longer
`inline` but it should not cause a significant regression
even if link-time optimization fails to inline it.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Flag: EXEMPT refactor
Change-Id: I2b2f2a06298984e5df1e725362383871fd3ba552