Add PaletteMapPriority
ART sometimes needs access to the mapping itself, so that it can,
for example, use an interpolated priority for system daemons.

aosp/3287177 contains the current version of the proposed client code.

Add PaletteClientTest for MapPriority.

This is in preparation for allowing ART to further manipulate thread
priorities. See go/thread-priorities-in-android and b/330444460 and
its many friends.

(Cherrypicked from aosp/3355710, which has more discussion)

Flag: NONE As discussed on aosp/3355710, impractical to flag
Bug: 389104950
Test: atest PaletteClientTest
Test: Treehugger
Change-Id: Ic7ba204b2d25cb82c9db01b21389b958ae527793