Check for the strings offset being zero
OffsetToPtr expects a non-zero offset which was leading to a crash.
Zero offsets are valid, as long as the size is also zero. This is
checked in `CheckHeader`

As a drive-by also add zero-checks for map_off_ (which is not allowed
to be 0) and data_off_ (which is).

Bug: 404728595
Test: m test-art-host-gtest-art_runtime_tests64
Flag: EXEMPT bugfix
Change-Id: I5dc7dd8824e8174633d88adcc07d3e37f183ebe6