Enable support for time-based concurrent GC triggering.
The new time-based concurrent GC triggering will take effect if
the HeapMemoryGcCostFactor runtime option is set to a non-zero value. It
causes time to be taken into accout in the decision for when to trigger
GC: the longer it has been since the last GC, the fewer bytes need to be
allocated before the next GC is triggered. This should improve
allocation of GC effort across processes on device, resulting in less GC
effort for more memory savings with appropriate choice of
HeapMemoryGcCostFactor.

Bug: 425665469
Test: Device boots. Run MPTS to see impact on memory and GC effort.
Flag: com.android.art.rw.flags.enable_time_based_gc_triggering
Change-Id: Id44f3d4551e36f694eb6f4a3821bc9d58277f6f3