Check for thread suspension in MonitorEnter
This should improve suspension latency while we are in a sched_yield
loop trying to acquire a monitor. Explicitly checking seems easier and
cleaner than entering a suspended state here. It's clearly safe, since
this function can already temporarily release the mutator lock.

This should greatly reduce suspension timeouts in the relatively rare
case that we are suspending a thread currently in the sched_yield loop.
We have seen this reported a few times.

Flag: EXEMPT bugfix
Bug: 406635928
Test: TreeHugger
Change-Id: If74d19fc146f5f863f56023334c5a53283dd5c00