Add long running method entry only if the string is not empty
In preperation for enabling the flag, add the long running method entry
only when the string is not empty. We don't want to change the ANR
reports even with the flag enabled. This lets us test the codegen
changes without waiting for the changes to the ANR pipeline that
understand the additional entry.

Bug: 352518093
Test: art/test.py
Flag: com.android.art.flags.always_enable_profile_code
Change-Id: I6c4d0116ade97b41cf67176b91b6e6515ecd3396