Report correctly the metrics for failure cases in Dexopter
In Dexopter.java we attempt to parse the dex2oat error message to
collect information about the status, exit code, and signal when
the dex2oat invocation fails, in order to report metrics to StatsD.
In doing so, we are using the Matcher#matches() method, which returns
true only if the regular expression matches the full string against
which the regex is evaluated - which is not correct in this case, as
the status, exit code, and signal returned from dex2oat represent only
a substring of the full error message.

The fix is simply to call Matcher#find() instead, which scans the input
sequence looking for the next subsequence that matches the pattern.

Bug: 327134025
Flag: EXEMPT bugfix
Test: atest ArtServiceTests
Change-Id: I2942f0581b9d6f7c91f1b49a74553880e313c241