Clean up and speed up `RegisterAllocationResolver`.
Pull some invariants out of a loop in `Resolve()` and
use another iterator type (without `PrefetchNext`).
Clean up for-loops.

Generalize `ZipCount()` to avoid an explicit call to the
`MakeIterationRange()` and rename the `MakeIterationRange`
parameter to make it clear it applies to any container-like
argument that provides `begin()` and `end()`. Remove the
`static` qualifier from some inline functions.

Remove `SsaLivenessAnalysis::number_of_ssa_values_` as the
size of `instructions_from_ssa_index_` provides the same
information. This provides context to some of the for-loop
rewrites in `RegisterAllocationResolver`.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Flag: EXEMPT refactor
Change-Id: I3e3ddae51a6f41d957c942d487f43a982ff1a6ed