Make SIGBUS handler accelerate suspend checks
Normally on arm64, we need to pass two suspend points for before a
suspend request results in a SIGSEGV and we notice the request.
The first suspend check copies the newly cleared suspend trigger
location into the dedicated register, and the second one tries to load
from the now zeroed address.

Immediately after a GC flip, this may take a long time, since a number
of normally fast memory accesses may result in page faults that cause
the CMC collector to compute the content of the referenced pages.

This reduces that latency somewhat by causing the CMC collector's signal
handler to copy cleared suspend trigger into the register, taking the
role of the first suspend point, and causing the next real suspend point
we encounter to actually trigger a SIGSEGV and check for a suspension
request.

This is hoped to reduce suspension latency immediately after a GC.
It should have near zero impact at other times, since we just check the
suspend trigger in the Thread structure.

This is ineffective if we are running interpreted code or are in the
runtime. But in those cases we normally detect suspend requests by
directly checking Thread flags, so we would do so at the first check in
any case.

Flag: NONE bug fix
Bug: 375053849
Test: Treehugger
Change-Id: I9d7490d260a35c6e866d948660c8739da57a11cd