Optimizing: Rewrite instruction visiting.
Improve performance by rewriting the instruction visiting in
a way that clang++ can optimize as an indexed virtual call.
This is better than dispatching via `Accept()` as we have
less branching and do not need to bring `Accept()` functions
scattered across the `libart.so` to instruction cache.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Bug: 413605257
Flag: EXEMPT refactor
Change-Id: I8e3f6fdd880a44679e4b96e35457a18cdf0bdb34