Add *_static_defaults for the simulator and use them.
libartd-simulator is an ART internal library, so it needs to be
statically linked into tests. It appears this was missed in
https://r.android.com/2623410.

Since tests link the simulator statically they need to bypass the
container that loads it with dlopen.

This CL adds the simulator static defaults to the generic test runtime
instead of the compiler tests specifically. That's necessary to avoid
an ASAN ODR error on host because libz gets both statically linked into
the test and loaded from libz-host.so by the LLVM bootstrap. This will
be addressed in a later CL where the libz linkage is fixed.

Since the simulator is currently only used in the compiler gtests, this
CL only adds *-for-test_static_defaults, but later it'll probably need
the corresponding *_static_defaults without "for-test" for the static
runtime.

Also add the proper runtime_libs dependencies to the container libs,
corresponding to the dlopen calls in them.

Test: art/tools/buildbot-build.sh
Test: art/test/testrunner/run_build_test_target.py -j64 art-gtest-asan
Bug: 347036796
Flag: TEST_ONLY
Change-Id: Iabcd577a3cbfa1bd98a43a70f79668b6198e4bdd