Add compiler argument support for SDK_INT evaluation
Add support for specifying assumed values for ART fields with a
new compiler argument of the form:
  * --assume-value=Landroid/os/Build$VERSION;->SDK_INT:$SDK_INT

Initially, we only support parsing SDK_INT, but this may be relaxed in
the feature to support arbitray fields (or even methods). Supply this
argument to the compiler for both AOT and JIT compilation, behind
a new trunk stable flag.

A follow-up change will add a constant folding pass that replaces static
field gets for assumed values with the provided values. Until that is
added, and the trunk stable flag is enabled, this should have no
functional impact on generated code.

Bug: 204924812
Test: presubmit
Test: atest --host art_artd_tests
Flag: com.android.art.flags.compile_sdk_int_constant
Change-Id: I162d9a0f1095db3b884d27c09c4a726494ba41e3