Speed up OatWriter by caching computed values
We have SafeMap (i.e. ordered maps) containers since we want a well
defined order for writing the oat file. However, this order becomes
slow when querying value when we don't care about the order. We can
speed up OatWriter by caching the results, and use that cache for
the times we want a value from a key, but we don't care about
iterating in said order.

Based on local compiles, this improves ~1.3-2.8% of compile time.

Bug: 393108375
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Flag: EXEMPT refactor
Change-Id: Ib6aa7b51182361ace669004e17a571cd0219f580