Priority fixes and minor cleanups
Rather than checking whether we managed to set a nonzero niceness
value, check for differences in the observed values. We are not
necessarily running at niceness zero.

Fix 051-thread so that it does not check that we set the correct
priority value in cases in which we cannot actually change Linux
niceness. This reduces test coverage, but that should be largely restored when we add the new palette call to directly retrieve the priority mapping.

Clarify that GetNativePriority() actually returns the native priority
and does not go through PaletteSchedGetPriority() anymore. Hence it is
no longer affected by palette_fake.

Remove the remaining use of PaletteSchedGetPriority().

Minor cleanups and kUseFakeOnHost debugging hook.

Bug: 413707588
Test: Manual
Flag: EXEMPT bug fix
Change-Id: Ib761ec65fb446c44a5f5a1663af67fd9317aba3d