Set priorities in decreasing order
GetPriorityMap should try to set the highest priority first when
determining the Java-priority-to-niceness mapping. This should cause it
to fail immediately if we cannot set the priority to higher values. This
potentially avoids a situation in which we lower the priority, and then
discover that we cannot undo that effect.

I am not aware of any current failures caused by this, since we avoid
running this code on host, where it could matter. But I would expect
someone to bump into this in the future.

Flag: EXEMPT bugfix
Test: Treehugger
Bug: 406536943
Change-Id: I946ac567024faad3eb9062b51057fd6262e11cc2