Handle OnStart timeouts gracefully
The current behaviour is that we require all tracing sessions with the
oome data source to successfully connect before we flush the data.
This is not ideal because a session might be configured to reject the
trigger based on the producer name (or timeout for any other reason). We
still want to flush to connected data sources.

Flag: EXEMPT bugfix
Test: atest HeapprofdJavaCtsTest, manual test CTS (concurrent with tracing session using producer filters)
Bug: 368365125
Change-Id: I999b81293d0992128fa1bfe47377fe11f5cc04ed