Change the behavior of `pm compile --reset`.
Since the `pm compile --reset` command was introduced in 2016, its
behavior has been documented as:

restore package to its post-install state

while its actual behavior on Android T- is to clear the profile data and
uncompile the app.

At the time where the command was introduced, "post-install state" ==
"uncompiled".

Over the years, the post-install behavior keeps changing (extract ->
interpret-only -> quicken -> speed-profile), but the doc and the
behavior of `pm compile --reset` never changed until Android U, so its
doc no longer reflected its behavior.

In Android U, we re-implemented dexopt in ART Service, and we believed
that "speed-profile" with the .dm file was the correct behavior to match
the doc, so we made it so.

However, in practice, developers' expectation seems to be based on its
actual behavior rather than how the doc describes it. Most developers
use `pm compile --reset` for clearing the profile data and uncompiling
the app. Our change in Android U to correct the behavior actually broke
developers.

Our decision is to put `pm compile --reset` back to the T- behavior
(i.e., uncompiling the app) and change the doc because this is the least
disruptive solution: `pm compile --reset` will just work in the way
developers expect, with no need for developers to change on their side.

Admittedly, this change introduces a behavior change of an existing
shell command through Mainline, which is not the best practice because
causes inconsistency between ART versions within the same API level
(e.g., API level 34 with the latest ART module will have the one
behavior, while the same API level with an older ART module has
another). However, we consider this change as a bug fix, so we believe
it is okay.

Flag: EXEMPT bugfix
Bug: 409506548
Test: atest ArtServiceTests
Test: adb shell pm compile --reset <pacakge-name>
Change-Id: If8fb13068b914b3728d28bcbe5f390d58b1bea9b