Update on pre-zygote native-roots of dirty dex-cache in compaction pause
We were visiting all pre-zygote fork native-roots in the compaction
pause. This meant that all ArtMethod/ArtField arrays of class and
dex-cache arrays allocated in zygote space will be visited. What we
really need to visit is dex-cache arrays, and that too only if the
corresponding card is dirty.

In this CL we identify the pre-zygote allocated native-root arrays of
dex-caches in immune spaces and only visit them if the dex-cache is
dirty.

Bug: 422561794
Test: device boot
Flag: EXEMPT refactor
Change-Id: I6eba305cd3e6a6471bc72e8ceb79c8aea1ea7a06