UFFD register entire moving space to recycle free pages in userspace
With COPY ioctl, we used to register only the used portion of the
moving-space during compaction to avoid userfaults on the unused portion
when new TLABs are allocated and accessed. This was necessary because we
couldn't map a writable page. The only mechanism we had was ZEROPAGE
ioctl which maps the shared zero-page at destination. That means
subsequent write would again cause a page-fault, at which point kernel
would allocate a private zeroed page.

Furthermore, in order to support this partial UFFD registration we had
to do some kernel specific tricks, which are not ideal.

With MOVE ioctl, we can recycle the free from-space pages directly from
userspace, saving us madvise effort, as well as unintended RSS increase
during times when CPU is contended and gc-thread doesn't get to release
memory as quickly as it should. Additionally, we greatly simplify the
UFFD registration process and make it more robust by first doing the
registration and then mremap, leaving no gaps for any random thread to
map some page on the moving space.

Test: local ART module install and reboot
Flag: EXEMPT refactor
Bug: 343220989
Bug: 422561794
Change-Id: I5ce5babb70cb24e5b8d069f96731d4ceff81930a