Check for no-reference case first thing in VisitReferences
Previously, VisitInstanceFieldsReferences(), which asserts that
reference bitmap should not be 0, was called on a record object
only if no other bit in class-flags was set. This meant that for
records with no references in them won't call the function
and therefore the assertion never failed.

With the recent class-flags changes in the MOVE ioctl CL, we ended up
slightly changing the behavior by calling the function on any record
object. In this CL we make sure to check no reference case first thing.

Flag: EXEMPT bugfix
Bug: 422497716
Bug: 343220989
Test: art/tools/run-libcore-tests.sh --mode=host --debug --gcstress
Change-Id: Ia4f9c505ea6fcfc6a377a9cf3bde6e516583776e