Don't use ScopedJitSuspend for redefinition.
ScopedJitSuspend waits for any ongoing Jit tasks to finish. It is not
safe to wait when servicing the redefinition request. Debugger might
have paused all the threads and the Jit thread might be waiting for one
of these threads to finish. So this would lead to a deadlock. So this CL
instead stops accepting any new tasks. We implement a mechanism to
discard any code for any ongoing compiles that finished after the
redefinition has finished.

For in-place updates we don't need to discard the Jited code. Hoewever,
it is not safe to allow a compilation for methods that are being
redefined. To keep the implementation simple, we just mark the
redefinition as structural when there are any ongoing compilations.

Bug: 298722987
Flag: EXEMPT Bugfix
Test: art/test.py

Change-Id: If9959fee61d409fe054a44e4d08ac2cbcfba6b5d