Optimizing: Refactor `Is{Entry,Exit}Block()`.
Move these functions to `HGraph` to avoid simple inline
functions in `HBasicBlock` depending on the definition of
`HGraph`, so that we can avoid including `graph.h` from
`basic_block.h` when we move the `HBasicBlock` to is own
files in a follow-up change.

One could make a case that this improves performance as we
replace some directly depenent loads with independent loads
but that improvement shall probably be lost in the noise.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Flag: EXEMPT refactor
Change-Id: I7e4875b5cbd7dec270dddda5da9e6d25ba29b90a