Allow small dynamic symbols.
The current algorithm removes a symbol if it's found in the dynamic
symbol table. The old code assumed that any dynamic symbol that's
in the regular symbol table has a size larger than or equal to the
regular symbol. This isn't always true, so now check the sizes and
do not remove a regular symbol if it's larger than the one in the
dynamic symbol table.

Bug: 424380322

Flag: EXEMPT not flaggable

Test: No longer aborts when running on shared libraries with smaller
Test: dynamic symbols.
Test: Verified the symbol is present in the compressed symbol table
Test: when the regular symbol is larger than the dynamic one.
Change-Id: I6f9c8916a45260d72bf1e9ab1bcd114ec91f3c55