Store DexFiles in the heap for the class verification fuzzer
We copy both the data, and the StandardDexFile container into the heap.
We do this to make sure that the DexFiles are accessible during GC.
Otherwise, we get ASAN crashes.

For this crash to manifest itself a GC needs to be run, and the dex
file has to have been collected already so we need to run at least two
DEX files before running the GC. Since our test infrastructure doesn't
support testing that, we don't have a regression test for this.
However, as it is a fuzzer failure (and not a real failure), it is fine
to merge without a regression test.

Bug: 426109304
Test: m test-art-host-gtest-art_runtime_tests64
Flag: EXEMPT bugfix
Change-Id: Ia7b9d778e0334f2457853b2be460275d0dac5e9a