build: Rename Android.mk to art.mk for SOONG_ONLY cleanup
This change is part of a multi-repo refactoring and depends on the
change in platform/build/core that updates `build/core/main.mk`.

As part of the effort to clean up Makefiles in preparation for
SOONG_ONLY builds, this change renames `art/Android.mk` to
`art/art.mk`.

This is a preparatory refactoring step. The `art.mk` file still contains
the necessary legacy Make logic for ART, but by renaming it, the build
system will no longer discover and include it automatically. Instead,
this file will now be explicitly included by `build/core/main.mk`.

Flag: EXEMPT Build infra and tools

Bug: 396022230
Test: lunch && m build-art-host build-art-target
Change-Id: I510544e1f8cece704cbdb4a398d58d8192c597ee