Bump up thread priority in more places
Remove the generic priority increase around reference processing and
replace it with broader increases around CMC and CC collector sections
that tend to block other threads. Also increase priority when collecting
because we're out of memory.

Tweak the ScopedPriorityChange API and implementation to make that
easier. Add a Reset() function, and no longer require the destructor to
be executed with the mutator lock held, provided the Reset() function is
explicitly called.

Some of these changes are a bit more invasive than desired, so that the
ScopedPriorityChange operations are called with the mutator lock held,
where required.

Limit inlining of ScopedPriorityChange operations to those pieces that
can be cheap.

The additional CHECKs in GetCachedNiceness pointed out that DumpState()
was not following the proper protocol for accessing the peer. Fixed
that, at the cost of very partially punting if we have a pending thread
flip.

Bug: 417492941
Bug: 431470883
Flag: com.android.libcore.niceness_apis
Test: Build and boot with extra logging
Change-Id: I61875f2ce0ead7ec2119c975da15225ad7087691