Use COPY ioctl for all threads when EBUSY returned from MOVE
Earlier when EBUSY was returned a gc-thread, we were would retry a
couple of times with MOVE. If it kept failing for few times, then we'd
fallback to COPY. That seems to be causing issues by making the
compaction phase sometimes be really long.

With this fix, we immediately use COPY ioctl regardless of which thread
gets EBUSY returned by MOVE ioctl. Additionally, we call madvise on the
src page for gc-thread to avoid rss increase.

Flag: EXEMPT bugfix
Bug: 430829352
Bug: 343220989
Test: crystalball infra
Change-Id: Ica777943f7534d86c40ce1fa438d3934468e1d4a