Optimizing: Avoid callee-save temp for @CriticalNative.
Except on x86 where we don't have a viable substitute.

This often results in generating less code by avoiding
spills. However, note that on arm and riscv64 we may
sometimes generate larger code as we may use 32-bit
encoding instead of 16-bit for one of the loads.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: run-gtests.sh
Test: testrunner.py --target --optimizing
Bug: 112189621
Flag: EXEMPT refactor
Change-Id: Ifc2c74ae3e15d98c35ea9444298fd066c24c3eca