Check enable_profile_code when accepting  boot images
Odrefresh does a few checks to see if boot images and other system
server jars are upto date. It wasn't required to check for the
oat header being valid. Since we now use a read-write flag to guard the
low-overhead tracing feature we have to check if these are compiled with
the tracing code or not. So this CL adds a check in odrefresh when
checking if the artifacts are upto date.

Bug: 352518093
Test: art/test.py
Test: 1. m droid
2. Flash device
3. adb shell aflags enable com.android.art.rw.flags.enable_profile_code_rw
4. adb reboot; adb wait-for-device; adb logcat -d
5. Manually verify from odrefresh logs that boot image and other system
   server artifacts were recompiled.
6. adb reboot; adb wait-for-device; adb logcat -d
7. Manually verify from odrefresh logs that boot image and system server
   artifacts were valid and were not recompiled.
Flag: com.android.art.rw.flags.enable_profile_code_rw

Change-Id: Icadc6ab66f4f209df47eb15a3702f268171fa0e3