Improve detection of files in ART and Conscrypt APEXes on host.
See comment for rationale. This change only affects the predicates used
in hiddenapi to detect the jars and native libs that are in the
core-platform domain. It's quite possible the reasoning extends to
other use cases and should be applied to more file_utils functions, but
that is out of scope here.

Test: art/test/run-test --host --64 --optimizing 674-hiddenapi
      with http://ag/34242953 which extends the test to check some
      core-platform APIs that shouldn't be accessible from platform
Test: atest art_standalone_libartbase_tests
Bug: 403305904
Flag: EXEMPT bugfix
Change-Id: I4862f5455574917e7d1b439ac5e60823be243da8