arm/arm64: Clean up visiting in arch simplifiers.
Do not replace users of the currently visited instruction.

Rewrite shifter operand merging and multiply accumulate
simplification to delay the replacement until we visit the
user. An `Add` may be eligible for both optimizations and
we shall only do the shifter operand merging in that case,
so we may actually miss a simple MultiplyAccumulate pattern
for its other input but that should be extremely rare.

Test: run-gtests.sh
Test: testrunner.py --target --optimizing
Bug: 181943478
Flag: EXEMPT refactor
Change-Id: I536e73c257b33c1dfea068553554c0b9d64d37a4