Check if clock_gettime returns a failure
We are seeing occasional failures when getting CLOCK_THREAD_CPUTIME_ID
time. This should only fail if either the timespec struct isn't valid or
if the operation isn't supported. I don't think these should happen but
it is good to verify that. So checking the return value and logging an
error would be good.

Also fixed a test to log a special error when the returned time is 0,
which is unexpected but valid. That will help with debugging
b/415170313.

Bug: 415170313
Test: art/test.py
Flag: EXEMPT Bug fix
Change-Id: Ia41f158854266d3820b1d2f52f32dbdcc87038dc