Refactor the code for reporting metrics in Dexopter
As the number of exceptional cases is growing, the code
responsible for reporting metrics around the dex2oat invocations
is getting more and more complicated and difficult to follow.

This CL refactors the reporting code to a session-based approach,
so that the session is modified across the dex2oat invocations
for different ABIs but the reporting code is only invoked once
in the finally block at the end of the session.

Bug: 327134025
Test: atest ArtServiceTests
Flag: EXEMPT refactor
Change-Id: Idd69022967c13ca57967b973cc0fb3665014d759