Maintain separate niceness and priority
We decided to keep independent priority and niceness values in Thread,
with niceness closer to the actual Linux niceness value, but not
inherited by Java threads, and not affecting Thread.getPriority().
This makes the necessary ART changes.

This is largely forced by the realization that
Process.setThreadPriority() values were not inherited, and clients
probably rely on that.

This is the ART part of that change.

Significantly expand priority test in 051-threads to actually test some
of this.

Remove JVM_SetThreadPriority(), which was unused, and violated an
expectation about ART setting OS niceness only to a previously cached
value.

Flag: EXEMPT bugfix
Bug: 417492941
Test: Treehugger
Change-Id: I584b49fd060d82680482e5762f4c11dec2f432f2