Update code sinking for ArraySets in the presence of loops
If both the ArraySet instruction and the index are within a loop,
then we can't sink the instruction as we risk not setting a value
in the array when we should have done it.

We can sink the ArraySet instruction if either the index or the
ArraySet instruction are out of the loop as either the same
"set" would be performed multiple times or the "set" would only
happen once anyway.

Note that for ArraySets even though we are using the array itself
to find a location, we are also checking in FindIdealPosition
that the found position is dominated by the ArraySet instruction.

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Bug: 405152615
Flag: EXEMPT bugfix
Change-Id: I887225189f8d959f67f7f14bede0c74c11121943