Fallback to system property if allow_profile_code isn't set
We used to default to setting allow_profile_code_ to false if it isn't
passed to dex2oat. This causes problems in test that don't pass this
flag but expect the boot images to be consistent. This cl changes the
implementation to either explicity specify allow_profile_code_ /
no_allow_profile_code_ when required. For other cases, we default to the
system property. This way tests work both when the flag is enabled or
disabled.

Bug:436662894, 352518093
Test: test.py with the rw flag enabled
Flag: com.android.art.rw.flags.enable_profile_code_rw

Change-Id: I240b3d54f19335d28b98e00b52f597ff2e6f6433