Use COPY ioctl in early stage of compaction to avoid lock contention
In the beginning of the compaction phase, it is expected that many
mutator threads (and gc-thread as well) cause userfaults simultaneously.
Concurrent invocations of MOVE ioctl for the same src vma (from-space)
is showing indications of kernel lock contention as all of them race
to exclusively acquire a read-write semaphore related to the vma.

By using COPY ioctl we can avoid that lock contention. This also makes
sense because at that early stage there are no free pages available in
the userspace to recycle. Therefore, the MOVE ioctl doesn't offer any
benefit in terms of limiting RSS increase too.

Flag: EXEMPT refactor
Test: manual
Change-Id: I8d68c7a52bb7a8a4928593b7cdbb8126a429f58b