ART: Clean up `InternTable::InternWeak()`.
Align `InternWeak()` overloads with the `InternStrong()` to
prepare for making `const-string` interns weak. One overload
is removed and callers in `ClassLinker` allocate the string
explicitly before calling `InternWeak(). One new overload is
added and mimics the corresponding `InternStrong()` overload
behavior to try avoiding the `String` allocation. The new
overload is not used in the ART code yet but it is covered
by additional code in `InternTableTest.Intern`.

Also remove one overload of `InternStrong()` and flag its
old use with `FIXME` because it should be `InternWeak()`.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 117621117
Flag: EXEMPT refactor
Change-Id: I4dfe2af833ae1dc5f373133ecb8f9cf915cb24c1