Adjust concurrent_start_bytes if gc-type is changing
We use bytes-allocated-during-gc as a measure of how much legroom we
would need to finish a background GC before hitting target-footprint.

However, we were not adjusting this if the gc-type is changing in
generational configuration. This means when we move from full-heap GC to
young GC, we would end up using a larger legroom than what is required,
triggering the young GC sooner.

This was revealed only after ag/34931978 because earlier
'kMaxConcurrentRemainingBytes' was almost always the chosen value.

Flag: EXEMPT bugfix
Bug: 437372741
Test: analyse perfetto traces of regressed crystalball tests
Change-Id: I59f6af0d37461a161a71565b1085652e01db7721