Inline and other optimizations in functions in marking fastpath
Also found some mistakes (like emitting the read-barrier even though we
don't use it in CMC) while working on this CL and are now fixed.

Running GC-intensive art-tests in gcstress doesn't show any clear
regression or improvement. But that could be due to unusual heap usage
in these tests as compared to actual apps. We typically either use
no-reference objects/arrays or object-arrays to fill up the heap.
Regular objects, which are common in real-world are not common in
art-tests.

Flag: EXEMPT refactor
Test: treehugger bloat buster
Change-Id: If8fe60fc57b32583cdf8ec23d4687ed30514c4f3