Madvise recyclable pages during compaction when using MOVE ioctl
We do so with COPY ioctl. But, initially chose to skip it with MOVE
as we already reuse pages for compacted pages. But the downside is that
we may hold on to too many free pages (than what we really need for
compaction) until the phase finishes, resulting in higher RSS/LMKs etc.

Also renamed GetFreePagesForMapping() to a more appropriate name and
updated a comment to correctly reflect why we have to skip madvise (in
case of COPY) sometimes.

Flag: com.android.art.rw.flags.use_uffd_move_ioctl_cmc_gc
Bug: 343220989
Test: crystalball infra
Change-Id: Ia420217c24a5f0d21d029b16b3280f438a3f3576