Keep Pre-reboot Dexopt staged files for a while if not committed.
If an OTA package has SWITCH_SLOT_ON_REBOOT=0, on reboot, the OTA is by
default not applied. In this case, we should keep the staged files for a
while.

Before this change, bg-dexopt always cleans up the staged files after a
reboot.

After this change, bg-dexopt keeps the staged files for a specific
number of days, specified by "dalvik.vm.pr_dexopt_retention" and
defaults to 30.

To achieve this, this change adds a Pre-reboot Dexopt staged metadata
file, whose mtime indicates when the staged files were created.

Bug: 436189791
Bug: 353430843
Test: atest art_standalone_artd_tests
Test: atest ArtServiceTests
Flag: EXEMPT BUGFIX
Change-Id: I8a61d3ec1a35d7493671c6dbf5cda2db448dcbac