Suppress FileNotFoundException in PreRebootStatsReporter.
When there is an OTA or Mainline update, Pre-reboot Dexopt writes
pre-reboot stats and persists it in /data/system/pre-reboot-stats.pb.
Later, this information is combined with after-reboot stats and reported
to statsd.

When background dexopt runs, it creates an AfterRebootSession to report
that it cleans up Pre-reboot Dexopt staged files if there's any.
However, in most of the cases, there's no stats to report, because
there's no OTA or Mainline update.

Before this change, when there's no stats for background dexopt to
report, AfterRebootSession logs a nasty stack trace because
/data/system/pre-reboot-stats.pb is not found.

After this change, it no longer logs a stack trace for this expected
case.

Bug: 436189791
Test: Presubmit
Flag: EXEMPT DEBUG
Change-Id: I1c05d1bf8f50ca4caebe7a9ec322041cc588ca32