Remove redundant timestamp fetch when recording exit events
In art_quick_record_long_running_exit_trace_event we already fetch the
timestamp counter before checking if the previous entry was an entry /
exit event. We don't have to re-fetch it. This was an oversight in an
earlier CL. It was initially implemented with one timestamp fetch in
each branch but it was optimized to move it to the common part, so that
there is no immediate dependency on the value.

Bug: 352518093
Test: art/test.py
Flag: com.android.art.rw.flags.enable_profile_code_rw

Change-Id: Ie8f88ff3169b169451308dc3672fbcbe534acd75